refactor: introduce request object for bulk secrets update
